Fixed https://github.com/adcade/saucey/issues/16, Updated Readme, added new dictionary definitions (tbc)
